+++ fs/lib/src/file_stream_sink.dart
@@ -1,0 +1,13 @@
+import 'dart:async';
+
+/// File streaming sink.
+abstract class FileStreamSink implements StreamSink<List<int>> {
+  /// Returns a [Future] that completes once all buffered data is accepted by the
+  /// underlying [StreamConsumer].
+  ///
+  /// This method must not be called while an [addStream] is incomplete.
+  ///
+  /// NOTE: This is not necessarily the same as the data being flushed by the
+  /// operating system.
+  Future<void> flush();
+}